+++ packages/graphql-server-hapi/src/hapiApollo.test.ts
@@ -2,7 +2,7 @@ import * as hapi from 'hapi';
 import { graphqlHapi, graphiqlHapi } from './hapiApollo';
 import 'mocha';
 
-import testSuite, { Schema, CreateAppOptions } from 'graphql-server-integration-testsuite';
+import testSuite, { schema as Schema, CreateAppOptions  } from 'graphql-server-integration-testsuite';
 
 function createApp(options: CreateAppOptions) {
   const server = new hapi.Server();
@@ -24,9 +24,7 @@ function createApp(options: CreateAppOptions) {
       register: graphiqlHapi,
       options: {
         path: '/graphiql',
-        graphiqlOptions: {
-          endpointURL: '/graphql',
-        },
+        graphiqlOptions: (options && options.graphiqlOptions) || { endpointURL: '/graphql' },
       },
   });